Explicitly cast slaveAddress to unsigned int
The QByteArray provides a signed char and casting it to int
retains the sign. The slave address is defined as unisgned int
though.

+++ src/serialbus/qmodbusrtuserialmaster_p.h
@@ -130,7 +130,7 @@ public:
             const QModbusResponse response(fcode,
                                            completeAduFrame.mid(2, completeAduFrame.size() - 4));
 
-            if (!canMatchRequestAndResponse(response, completeAduFrame.at(0))) {
+            if (!canMatchRequestAndResponse(response, quint8(completeAduFrame.at(0)))) {
                 qCWarning(QT_MODBUS) << "Cannot match response with open request. "
                                         "Ignoring response.";
                 return;
